Rename cfe_simulation.py -> cf_simulation.py
